requirements.txt
nextcord==2.4.2
python-dotenv==1.0.0
termcolor==2.3.0
